added an error handler for 404 errors

+++ app.py
@@ -1,4 +1,4 @@
-from flask import Flask, jsonify, request, abort, redirect
+from flask import Flask, jsonify, request, abort, redirect, make_response
 from sqlalchemy import create_engine, Column, Integer, String
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.ext.declarative import declarative_base
@@ -60,33 +60,17 @@ def get_companies():
 
 @app.route(BASE_DIR + "/<company_name>", methods=["GET"])
 def get_company(company_name):
-    company = session.query(Company).filter_by(name=company_name).one()
-    return jsonify({"company": company.to_dict()})
+    company = session.query(Company).filter_by(name=company_name)
+
+    if(company.count() == 0):
+        abort(404)
+
+    return jsonify({"company": company.one().to_dict()})
 
 
 @app.route(BASE_DIR, methods=["POST"])
 def add_company():
-    if not request.get_json() or "name" not in request.get_json():
-        abort(400)
-
-    if("name" in request.get_json()
-            and type(request.get_json()["name"]) != str):
-        abort(400)
-
-    if("employees" in request.get_json()
-            and type(request.get_json()["employees"]) != int):
-        abort(400)
-
-    if("location" in request.get_json()
-            and type(request.get_json()["location"]) != str):
-        abort(400)
-
-    if("email" in request.get_json()
-            and type(request.get_json()["email"]) != str):
-        abort(400)
-
-    if("industry" in request.get_json()
-            and type(request.get_json()["industry"]) != str):
+    if(not request_valid() or "name" not in request.get_json()):
         abort(400)
 
     company_name = request.get_json().get("name")
@@ -110,30 +94,14 @@ def add_company():
 
 @app.route(BASE_DIR + "/<company_name>", methods=["PUT"])
 def update_company(company_name):
-    if(not request.get_json()):
-        abort(400)
-
-    if("name" in request.get_json()
-            and type(request.get_json()["name"]) != str):
-        abort(400)
-
-    if("employees" in request.get_json()
-            and type(request.get_json()["employees"]) != int):
-        abort(400)
-
-    if("location" in request.get_json()
-            and type(request.get_json()["location"]) != str):
-        abort(400)
-
-    if("email" in request.get_json()
-            and type(request.get_json()["email"]) != str):
-        abort(400)
-
-    if("industry" in request.get_json()
-            and type(request.get_json()["industry"]) != str):
+    if(not request_valid()):
         abort(400)
 
     company = session.query(Company).filter_by(name=company_name)
+
+    if(company.count() == 0):
+        abort(404)
+
     company.update(request.get_json())
     company = company.one()
     session.commit()
@@ -144,10 +112,46 @@ def update_company(company_name):
 @app.route(BASE_DIR + "/<company_name>", methods=["DELETE"])
 def remove_company(company_name):
     company = session.query(Company).filter_by(name=company_name)
+
+    if(company.count() == 0):
+        abort(404)
+
     company.delete()
     session.commit()
 
     return jsonify({"result": True})
+
+
+@app.errorhandler(404)
+def not_found(error):
+	return make_response(jsonify({"error": "Not found"}), 404)
+
+
+def request_valid():
+    if not request.get_json():
+        return False
+
+    if("name" in request.get_json()
+            and type(request.get_json()["name"]) != str):
+        return False
+
+    if("employees" in request.get_json()
+            and type(request.get_json()["employees"]) != int):
+        return False
+
+    if("location" in request.get_json()
+            and type(request.get_json()["location"]) != str):
+        return False
+
+    if("email" in request.get_json()
+            and type(request.get_json()["email"]) != str):
+        return False
+
+    if("industry" in request.get_json()
+            and type(request.get_json()["industry"]) != str):
+        return False
+
+    return True
 
 
 if(__name__ == "__main__"):
